add sql boolean

+++ cog/note.py
@@ -21,13 +21,14 @@ class Notes(cog_extension):
             TIMEE TIMESTAMP,
             PRIVATE BOOL);
             ''')
+        self.sql = False
         self.con.commit()
         self.note_list = []
         self.time = []
         notes = self.cur.execute(
             "SELECT * FROM `NOTE`  ORDER BY `TIMEE` DESC")
         for a in notes.fetchall():
-            self.time.append(int(a[3]))
+            self.time.append(int(a[2]))
             self.note_list.append(a)
 
         async def nowtime():
@@ -37,7 +38,7 @@ class Notes(cog_extension):
                 print(time)
                 self.time.sort()
                 if time > self.time[0]:
-                    await self.channel.send(f'Hey! {self.bot.get_user(int(self.note_list[0][0]))} {self.note_list[0][1]}')
+                    await self.channel.send(f'Hey! {self.bot.get_user(int(self.note_list[0][0])).mention} {self.note_list[0][1]}')
                 await asyncio.sleep(300)
 
         self.timesssss = self.bot.loop.create_task(nowtime())
@@ -56,13 +57,22 @@ class Notes(cog_extension):
         if len(time) != 12:
             await ctx.send('Please follow the style yyyymmddhhmm ex. 202201190857')
             return 'Error!'
-        if int(time):
-            self.cur.execute(
-                "INSERT INTO `NOTE` (`USER`,`MESSAGE`,`TIMEE`,`PRIVATE`) VALUES(?,?,?,?)", (author, name, time, private))
-            # self.con.commit()
-
+        if self.sql == False:
+            if int(time):
+                self.sql = True
+                self.cur.execute(
+                    "INSERT INTO `NOTE` (`USER`,`MESSAGE`,`TIMEE`,`PRIVATE`) VALUES(?,?,?,?)", (author, name, time, private))
+                self.con.commit()
             await ctx.send(content='test', delete_after=5.0)
             await ctx.send(f'**{name}** create success!')
+            notes = self.cur.execute(
+                "SELECT * FROM `NOTE`  ORDER BY `TIMEE` DESC")
+            for a in notes.fetchall():
+                self.time.append(int(a[2]))
+                self.note_list.append(a)
+            self.sql = False
+        else:
+            ctx.send('Wait for the last query finish!')
 
     @commands.command()
     async def delete(self, ctx, name):
@@ -78,9 +88,15 @@ class Notes(cog_extension):
                 await ctx.send('private message!')
                 return 'Falied!'
         else:
-            self.cur.execute("DELETE FROM NOTE WHERE MESSAGE=(?)", (name,))
-            # self.con.commit()
-        await ctx.send('Delete success!')
+            if self.sql == False:
+                self.sql = True
+                self.cur.execute("DELETE FROM NOTE WHERE MESSAGE=(?)", (name,))
+                self.con.commit()
+                self.sql = False
+                self.note_list.remove(name)
+                await ctx.send('Delete success!')
+            else:
+                ctx.send('Wait for the last query finish!')
 
 
 async def setup(bot):
